+++ buildSrc/src/main/kotlin/house/greenhouse/bovinesandbuttercups/gradle/Versions.kt
@@ -1,7 +1,7 @@
 package house.greenhouse.bovinesandbuttercups.gradle
 
 object Versions {
-    const val MOD = "2.2.10"
+    const val MOD = "2.2.11"
 
     // Minecraft
     const val MINECRAFT = "1.21.1"